+++ Sublime Text 3/Packages/SublimeHaskell/parseoutput.py
@@ -1,4 +1,7 @@
+# -*- coding: UTF-8 -*-
+
 import os
+import os.path
 import re
 import sublime
 import sublime_plugin
@@ -10,11 +13,11 @@ from collections import defaultdict
 PyV3 = version[0] == "3"
 
 if int(sublime.version()) < 3000:
-    from sublime_haskell_common import log, are_paths_equal, call_and_wait, get_setting_async, show_status_message_process, show_status_message
+    from sublime_haskell_common import *
 else:
-    from SublimeHaskell.sublime_haskell_common import log, are_paths_equal, call_and_wait, get_setting_async, show_status_message_process, show_status_message
+    from SublimeHaskell.sublime_haskell_common import *
 
-ERROR_PANEL_NAME = 'haskell_error_checker'
+ERROR_PANEL_NAME = 'haskell_error_panel'
 
 # This regex matches an unindented line, followed by zero or more
 # indented, non-empty lines.
@@ -27,11 +30,16 @@ output_regex = re.compile(
 # Extract the filename, line, column, and description from an error message:
 result_file_regex = r'^(\S*?): line (\d+), column (\d+):$'
 
+# Extract the filename, line, column from symbol info
+symbol_file_regex = r'^Defined at: (.*):(\d+):(\d+)$'
 
 # Global list of errors. Used e.g. for jumping to the next one.
 # Properly assigned being a defaultdict in clear_error_marks().
-# Structure: ERRORS[filename][m.line] = OutputMessage()
+# Structure: ERRORS[filename][m.start.line] = OutputMessage()
 ERRORS = {}
+
+# Global ref to view with errors
+error_view = None
 
 
 def filename_of_path(p):
@@ -40,42 +48,59 @@ def filename_of_path(p):
     # we have forward or backslashes on Windows.
     return re.match(r'(.*[/\\])?(.*)', p).groups()[1]
 
+class OutputPoint(object):
+    def __init__(self, line, column):
+        self.line = int(line)
+        self.column = int(column)
+
+    def __unicode__(self):
+        return u"{0}:{1}".format(self.line, self.column)
+
+    def __str__(self):
+        return self.__unicode__()
+
+    def __eq__(self, other):
+        return self.line == other.line and self.column == other.column
+
+    def to_point_of_view(self, view):
+        return view.text_point(self.line, self.column)
 
 class OutputMessage(object):
     "Describe an error or warning message produced by GHC."
-    def __init__(self, filename, line, column, message, level):
+    def __init__(self, filename, start, end, message, level):
         self.filename = filename
-        self.line = int(line)
-        self.column = int(column)
+        self.start = start
+        self.end = end
         self.message = message.replace(os.linesep, "\n")
         self.level = level
 
     def __unicode__(self):
         # must match result_file_regex
+        # TODO: Columns must be recalculated, such that one tab is of tab_size length
+        # We can do this for opened views, but how to do this for files, that are not open?
         return u'{0}: line {1}, column {2}:\n  {3}'.format(
             self.filename,
-            self.line,
-            self.column,
+            self.start.line + 1,
+            self.start.column + 1,
             self.message)
 
     def __str__(self):
         return self.__unicode__()
 
     def __repr__(self):
-        return '<OutputMessage {0}:{1}:{2}: {3}>'.format(
+        return '<OutputMessage {0}:{1}-{2}: {3}>'.format(
             filename_of_path(self.filename),
-            self.line,
-            self.column,
+            self.start.__repr__(),
+            self.end.__repr__(),
             self.message[:10] + '..')
 
-    def find_region_in_view(self, view):
+    def to_region_in_view(self, view):
         "Return the Region referred to by this error message."
         # Convert line and column count to zero-based indices:
-        point = view.text_point(self.line - 1, 0)
-        # Return the whole line:
-        region = view.line(point)
-        region = trim_region(view, region)
-        return region
+        if self.start == self.end: # trimmed full line
+            return trim_region(view, view.line(self.start.to_point_of_view(view)))
+        return sublime.Region(self.start.to_point_of_view(view), self.end.to_point_of_view(view))
+
 
 
 def clear_error_marks():
@@ -91,7 +116,7 @@ def set_global_error_messages(messages):
     clear_error_marks()
 
     for m in messages:
-        ERRORS[m.filename][m.line].append(m)
+        ERRORS[m.filename][m.start.line].append(m)
 
 
 def run_build_thread(view, cabal_project_dir, msg, cmd, on_done):
@@ -121,14 +146,24 @@ def wait_for_chain_to_complete(view, cabal_project_dir, msg, cmds, on_done):
     sublime.set_timeout(lambda: hide_output(view), 0)
 
     # run and wait commands, fail on first fail
+    stdout = ''
+    stderr = ''
+    output_log = output_panel(view.window(), '', panel_name = 'sublime_haskell_build_log', show_panel = get_setting_async('show_output_window'))
     for cmd in cmds:
-        exit_code, stdout, stderr = call_and_wait(
-            cmd,
-            cwd=cabal_project_dir)
-        if exit_code != 0:
-            break
+        output_text(output_log, ' '.join(cmd) + '...\n')
 
-    errmsg = stderr if stderr else stdout
+        cmd_p = call_and_wait(cmd, cwd = cabal_project_dir, wait = False)
+        lines = []
+        for cmd_line in cmd_p.stdout:
+            line = crlf2lf(decode_bytes(cmd_line))
+            lines.append(line)
+            output_text(output_log, line)
+        exit_code = cmd_p.wait()
+        stdout = '\n'.join(lines)
+        stderr = crlf2lf(decode_bytes(cmd_p.stderr.read()))
+    hide_panel(view.window(), panel_name = 'sublime_haskell_build_log')
+
+    errmsg = stderr
 
     # Notify UI thread that commands are done
     sublime.set_timeout(on_done, 0)
@@ -138,10 +173,19 @@ def wait_for_chain_to_complete(view, cabal_project_dir, msg, cmds, on_done):
 
 def format_output_messages(messages):
     """Formats list of messages"""
+    summary = {'error': 0, 'warning': 0, 'hint': 0}
+    for m in messages:
+        summary[m.level] = summary[m.level] + 1
+    summary_line = 'Errors: {0}, Warnings: {1}, Hints: {2}'.format(
+        summary['error'],
+        summary['warning'],
+        summary['hint'])
     if PyV3:
-        return '\n'.join(str(x) for x in messages)
+        details = '\n'.join(str(x) for x in messages)
+        return '{0}\n\n{1}'.format(summary_line, details) if details else ''
     else:
-        return u'\n'.join(unicode(x) for x in messages)
+        details = u'\n'.join(unicode(x) for x in messages)
+        return u'{0}\n\n{1}'.format(summary_line, details) if details else u''
 
 def show_output_result_text(view, msg, text, exit_code, base_dir):
     """Shows text (formatted messages) in output with build result"""
@@ -154,7 +198,7 @@ def show_output_result_text(view, msg, text, exit_code, base_dir):
     show_status_message_process(msg, success)
     # Show panel if there is any text to show (without the part that we add)
     if text:
-        if get_setting_async('show_output_window'):
+        if get_setting_async('show_error_window'):
             sublime.set_timeout(lambda: write_output(view, output, base_dir), 0)
 
 
@@ -166,7 +210,7 @@ def parse_output_messages_and_show(view, msg, base_dir, exit_code, stderr):
     # stderr = stderr.decode('utf-8')
 
     # The process has terminated; parse and display the output:
-    parsed_messages = parse_output_messages(base_dir, stderr)
+    parsed_messages = parse_output_messages(view, base_dir, stderr)
     # The unparseable part (for other errors)
     unparsable = output_regex.sub('', stderr).strip()
 
@@ -200,25 +244,25 @@ def mark_messages_in_views(errors):
             if view_filename is None:
                 continue
             errors_in_view = list(filter(
-                lambda x: are_paths_equal(view_filename, x.filename),
+                lambda x: os.path.samefile(view_filename, x.filename),
                 errors))
             mark_messages_in_view(errors_in_view, v)
     end_time = time.clock()
     log('total time to mark {0} diagnostics: {1} seconds'.format(
-        len(errors), end_time - begin_time))
+        len(errors), end_time - begin_time), log_debug)
 
 message_levels = {
     'hint': {
-        'style': 'comment.warning',
-        'icon': 'light_x_bright'
+        'style': 'sublimehaskell.mark.hint',
+        'icon': 'haskell-hint.png'
     },
     'warning': {
-        'style': 'comment.warning',
-        'icon': 'grey_x_light_shadow'
+        'style': 'sublimehaskell.mark.warning',
+        'icon': 'haskell-warning.png'
     },
     'error': {
-        'style': 'invalid',
-        'icon': 'grey_x'
+        'style': 'sublimehaskell.mark.error',
+        'icon': 'haskell-error.png'
     }
 }
 
@@ -226,32 +270,57 @@ message_levels = {
 # These next and previous commands were shamelessly copied
 # from the great SublimeClang plugin.
 
-class SublimeHaskellNextError(sublime_plugin.TextCommand):
+def goto_error(view, filename, line, column):
+    global error_view
+    if error_view:
+        show_output(view)
+        # error_region = error_view.find('{0}: line {1}, column \\d+:(\\n\\s+.*)*'.format(re.escape(filename), line), 0)
+        error_region = error_view.find('{0}: line {1}, column {2}:(\\n\\s+.*)*'.format(re.escape(filename), line, column), 0)
+        error_view.add_regions("current_error", [error_region], 'string', 'dot', sublime.HIDDEN)
+        error_view.show(error_region.a)
+    view.window().open_file("{0}:{1}:{2}".format(filename, line, column), sublime.ENCODED_POSITION)
+
+def get_next_value(v, lst, cycle = True):
+    # Get next value from list
+    if v is None:
+        if len(lst) > 0:
+            return lst[0]
+        return None
+    for x in filter(lambda k: k > v, lst):
+        return x
+    if cycle and len(lst) > 0:
+        return lst[0]
+    return None
+
+class SublimeHaskellNextError(SublimeHaskellTextCommand):
     def run(self, edit):
-        log("SublimeHaskellNextError")
         v = self.view
-        fn = v.file_name().encode("utf-8")
+        fn = v.file_name()
         line, column = v.rowcol(v.sel()[0].a)
-        line += 1
-        gotoline = -1
+        # line += 1
+        gotoline = None
+        gotocolumn = None
         if fn in ERRORS:
-            for errLine in sorted(ERRORS[fn].keys()):
-                if errLine > line:
-                    gotoline = errLine
-                    break
-            # No next line: Wrap around if possible
-            if gotoline == -1 and len(ERRORS[fn]) > 0:
-                gotoline = sorted(ERRORS[fn].keys())[0]
-        if gotoline != -1:
-            v.window().open_file("%s:%d" % (fn, gotoline), sublime.ENCODED_POSITION)
-        else:
-            sublime.status_message("No more errors or warnings!")
+            if line in ERRORS[fn]: # on some line, check if there are another error on same linee
+                gotoline = line
+                gotocolumn = get_next_value(column, sorted([e.start.column for e in ERRORS[fn][gotoline]]), cycle = False)
+                if gotocolumn is not None: # next error on same line
+                    goto_error(v, fn, gotoline + 1, gotocolumn + 1)
+                    return
+            # no error on this line, find next (and cycle through)
+            gotoline = get_next_value(line, sorted(ERRORS[fn].keys()))
+            if gotoline is not None:
+                # go to first error on line
+                gotocolumn = get_next_value(None, sorted([e.start.column for e in ERRORS[fn][gotoline]]), cycle = False)
+                if gotocolumn is not None: # found some
+                    goto_error(v, fn, gotoline + 1, gotocolumn + 1)
+                    return
+            show_status_message('No more errors or warnings!', priority = 5)
 
-
-class SublimeHaskellPreviousError(sublime_plugin.TextCommand):
+class SublimeHaskellPreviousError(SublimeHaskellTextCommand):
     def run(self, edit):
         v = self.view
-        fn = v.file_name().encode("utf-8")
+        fn = v.file_name()
         line, column = v.rowcol(v.sel()[0].a)
         line += 1
         gotoline = -1
@@ -264,7 +333,7 @@ class SublimeHaskellPreviousError(sublime_plugin.TextCommand):
             if gotoline == -1 and len(ERRORS[fn]) > 0:
                 gotoline = sorted(ERRORS[fn].keys())[-1]
         if gotoline != -1:
-            v.window().open_file("%s:%d" % (fn, gotoline), sublime.ENCODED_POSITION)
+            goto_error(v, fn, gotoline)
         else:
             sublime.status_message("No more errors or warnings!")
 
@@ -273,6 +342,12 @@ class SublimeHaskellPreviousError(sublime_plugin.TextCommand):
 def region_key(name):
     return 'subhs-{0}s'.format(name)
 
+def get_icon(png):
+    return "/".join([
+        "Packages",
+        os.path.basename(os.path.dirname(__file__)),
+        "Icons",
+        png])
 
 def mark_messages_in_view(messages, view):
     # Regions by level
@@ -281,7 +356,7 @@ def mark_messages_in_view(messages, view):
         regions[k] = []
 
     for m in messages:
-        regions[m.level].append(m.find_region_in_view(view))
+        regions[m.level].append(m.to_region_in_view(view))
 
     for nm, lev in message_levels.items():
         view.erase_regions(region_key(nm))
@@ -289,44 +364,73 @@ def mark_messages_in_view(messages, view):
             region_key(nm),
             regions[nm],
             lev['style'],
-            lev['icon'],
+            get_icon(lev['icon']),
             sublime.DRAW_OUTLINED)
 
 
-def write_output(view, text, cabal_project_dir):
+def write_panel(window, text, panel_name = "sublime_haskell_panel", syntax = None):
+    info_view = output_panel(window, text, panel_name = panel_name, syntax = syntax)
+    info_view.settings().set("result_file_regex", symbol_file_regex)
+    return info_view
+
+def write_output(view, text, cabal_project_dir, show_panel = True):
     "Write text to Sublime's output panel."
-    output_view = view.window().get_output_panel(ERROR_PANEL_NAME)
-    output_view.set("color_scheme", "Packages/User/SublimeLinter/Solarized (Dark) (SL).tmTheme")
-    output_view.set_read_only(True)
-    # Configure Sublime's error message parsing:
-    output_view.settings().set("result_file_regex", result_file_regex)
-    output_view.settings().set("result_base_dir", cabal_project_dir)
-    # Write to the output buffer:
-    output_view.run_command('sublime_haskell_output_text', {
-        'text': text })
-    # Set the selection to the beginning of the view so that "next result" works:
-    output_view.sel().clear()
-    output_view.sel().add(sublime.Region(0))
-    output_view.set_read_only(True)
-    # Show the results panel:
-    view.window().run_command('show_panel', {'panel': 'output.' + ERROR_PANEL_NAME})
+    global error_view
+    error_view = output_panel(view.window(), text, panel_name = ERROR_PANEL_NAME, syntax = 'HaskellOutputPanel', show_panel = show_panel)
+    error_view.settings().set("result_file_regex", result_file_regex)
+    error_view.settings().set("result_base_dir", cabal_project_dir)
 
 
-def hide_output(view):
-    view.window().run_command('hide_panel', {'panel': 'output.' + ERROR_PANEL_NAME})
+def hide_output(view, panel_name = ERROR_PANEL_NAME):
+    view.window().run_command('hide_panel', {'panel': 'output.' + panel_name})
 
+def show_output(view, panel_name = ERROR_PANEL_NAME):
+    view.window().run_command('show_panel', {'panel': 'output.' + panel_name})
 
-def parse_output_messages(base_dir, text):
+def tabs_offset(view, point):
+    """
+    Returns count of '\t' before point in line multiplied by 7
+    8 is size of type as supposed by ghc-mod, to every '\t' will add 7 to column
+    Subtract this value to get sublime column by ghc-mod column, add to get ghc-mod column by sublime column
+    """
+    cur_line = view.substr(view.line(point))
+    return len(list(filter(lambda ch: ch == '\t', cur_line))) * 7
+
+def sublime_column_to_ghc_column(view, line, column):
+    """
+    Convert sublime zero-based column to ghc-mod column (where tab is 8 length)
+    """
+    return column + tabs_offset(view, view.text_point(line, column)) + 1
+
+def ghc_column_to_sublime_column(view, line, column):
+    """
+    Convert ghc-mod column to sublime zero-based column
+    """
+    cur_line = view.substr(view.line(view.text_point(line - 1, 0)))
+    col = 1
+    real_col = 0
+    for c in cur_line:
+        if col >= column:
+            return real_col
+        col += (8 if c == '\t' else 1)
+        real_col += 1
+    return real_col
+
+def parse_output_messages(view, base_dir, text):
     "Parse text into a list of OutputMessage objects."
     matches = output_regex.finditer(text)
 
     def to_error(m):
         filename, line, column, messy_details = m.groups()
+        line, column = int(line), int(column)
+
+        column = ghc_column_to_sublime_column(view, line, column)
+        line = line - 1
         return OutputMessage(
             # Record the absolute, normalized path.
             os.path.normpath(os.path.join(base_dir, filename)),
-            line,
-            column,
+            OutputPoint(line, column),
+            OutputPoint(line, column),
             messy_details.strip(),
             'warning' if 'warning' in messy_details.lower() else 'error')